spark-batch/python_requirements.txt
backcall==0.1.0
bleach==2.1.3
blinker==1.3
cassandra-driver==3.14.0
certifi==2018.4.16
chardet==3.0.4
cloud-init==0.7.9
command-not-found==0.3
configobj==5.0.6
cryptography==1.2.3
cycler==0.10.0
decorator==4.3.0
entrypoints==0.2.3
html5lib==1.0.1
idna==2.7
ipykernel==4.8.2
ipython==6.4.0
ipython-genutils==0.2.0
ipywidgets==7.2.1
jedi==0.12.1
Jinja2==2.10
jsonpatch==1.10
jsonpointer==1.9
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.3
jupyter-console==5.2.0
jupyter-core==4.4.0
kaggle==1.3.12
kiwisolver==1.0.1
language-selector==0.1
MarkupSafe==1.0
matplotlib==2.2.2
mistune==0.8.3
nbconvert==5.3.1
nbformat==4.4.0
notebook==5.6.0
numpy==1.14.3
oauthlib==1.0.3
pandas==0.23.3
pandocfilters==1.4.2
parso==0.3.1
pexpect==4.6.0
pickleshare==0.7.4
powerline-status==2.3
prettytable==0.7.2
prometheus-client==0.3.0
prompt-toolkit==1.0.15
psutil==3.4.2
ptyprocess==0.6.0
pyasn1==0.1.9
pycurl==7.43.0
Pygments==2.2.0
pygobject==3.20.0
PyJWT==1.3.0
pyparsing==2.2.0
pyserial==3.0.1
python-apt==1.1.0b1
python-dateutil==2.7.2
python-debian==0.1.27
python-systemd==231
pytz==2018.5
PyYAML==3.11
pyzmq==17.1.0
qtconsole==4.3.1
requests==2.19.1
scikit-learn==0.19.1
scipy==1.0.1
Send2Trash==1.5.0
simplegeneric==0.8.1
six==1.11.0
ssh-import-id==5.5
terminado==0.8.1
testpath==0.3.1
tornado==5.1
tqdm==4.23.4
traitlets==4.3.2
ufw==0.35
unattended-upgrades==0.1
urllib3==1.22
virtualenv==15.2.0
wcwidth==0.1.7
webencodings==0.5.1
widgetsnbextension==3.2.1
xgboost==0.71
